Moved the HardwareLocality into its own set of files.
