+++ src/hw_locality.hpp
@@ -1,0 +1,60 @@
+/*
+ * Copyright (c) 2019, The Bifrost Authors. All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ * * Redistributions of source code must retain the above copyright
+ *   notice, this list of conditions and the following disclaimer.
+ * * Redistributions in binary form must reproduce the above copyright
+ *   notice, this list of conditions and the following disclaimer in the
+ *   documentation and/or other materials provided with the distribution.
+ * * Neither the name of The Bifrost Authors nor the names of its
+ *   contributors may be used to endorse or promote products derived
+ *   from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS ``AS IS'' AND ANY
+ * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
+ * PURPOSE ARE DISCLAIMED.  IN NO EVENT SHALL THE COPYRIGHT OWNER OR
+ * CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+ * EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+ * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+ * PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
+ * OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+#include <bifrost/affinity.h>
+
+#if BF_HWLOC_ENABLED
+#include <hwloc.h>
+class HardwareLocality {
+    hwloc_topology_t _topo;
+    HardwareLocality(HardwareLocality const&);
+    HardwareLocality& operator=(HardwareLocality const&);
+public:
+    HardwareLocality() {
+        hwloc_topology_init(&_topo);
+        hwloc_topology_load(_topo);
+    }
+    ~HardwareLocality() {
+        hwloc_topology_destroy(_topo);
+    }
+    int bind_memory_to_core(int core);
+};
+#endif // BF_HWLOC_ENABLED
+
+class BoundThread {
+#if BF_HWLOC_ENABLED
+    HardwareLocality _hwloc;
+#endif
+public:
+    BoundThread(int core) {
+        bfAffinitySetCore(core);
+#if BF_HWLOC_ENABLED
+        assert(_hwloc.bind_memory_to_core(core) == 0);
+#endif
+    }
+};
